Create is the views folder and the views files are stored in this folder. And Minor the change of forms
